File conversions can now be deleted. Overwrote write, create and unlink methods on  the parent class (oks.intranet.document) to enable automatic file conversion and retrieval on all its childs.
File conversion is automatically performed on write and create.
Converted files no longer needed are removed too. When a whole
record is dropped the whole folder is deleted.

Temporary files written to disk still need to be deleted. And there
are probably ways to improve efficiency or decrease db size